fix: move TrainerArgs, TrainerConfig into new trainer.config module
Avoids errors due to circular dependencies.
