+++ tests/test_train_gan.py
@@ -8,8 +8,7 @@ from torch.utils.data import DataLoader
 from torchvision import transforms
 from torchvision.datasets import MNIST
 
-from trainer import TrainerConfig, TrainerModel
-from trainer.trainer import Trainer, TrainerArgs
+from trainer import Trainer, TrainerArgs, TrainerConfig, TrainerModel
 
 is_cuda = torch.cuda.is_available()
