+++ trainer/__init__.py
@@ -1,5 +1,6 @@
 import importlib.metadata
 
+from trainer.config import TrainerArgs, TrainerConfig
 from trainer.model import *
 from trainer.trainer import *
